refactor: move GradleSpecifier into model
